refactor: extract duplicated I/O and metric helpers in scripts/ into _utils.py
Move repeated repo_path, rel_path, load_yaml/load_json, write_json/stable_json,
append_jsonl, json_hash, utc_now, git_output/git_dirty, fmt_rate, and the
metric_snapshot/metric_block pair into scripts/_utils.py. metric_snapshot now
takes include_query_type to absorb both variants.

Importers use the scripts-dir-on-path + bare `from _utils import ...` pattern,
matching the existing scripts/_eval_delta.py convention.

display_path, fmt_latency, fmt_delta, and run_harness.py's local
metric_snapshot(Path) are intentionally left in place — they look similar but
have different return contracts.

No behavior change in the RAG pipeline, ingestion, eval, or answer contract.
pytest -q: 469 passed.

Closes #254
